requirements-local.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
requests
python-dotenv
